+++ tests/test_client_cleanup.py
@@ -1,0 +1,7 @@
+import pysvn
+import pytest
+
+svn = pysvn.Client(repository_dir='./tests/test_svn')
+
+def test_cleanup():
+    svn.cleanup()